+++ CHANGELOG.md
@@ -7,8 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - ACF Pro can now be updated via `composer` as well as the legacy method
 - Added `--skip-git` for projects that are focused only on client communication
 - Added ability to turn off RSS feed per project, even if set globally
+- All plugins are activated when using `--prepare-with-reset` on a Wordpress project
 
-### [Fixed]
+### Fixed
 - Fixed success color rendering in default html theme
 - Fixed a bug in preparing new Wordpress projects
 - Fixed an activity log email formatting error